Refactor: Extract HTTP communication into DebuggerApiService
- Created Services folder with IDebuggerApiService interface
- Moved all background service HTTP calls from McpProxy to DebuggerApiService
- Updated DI container to register HttpClient and DebuggerApiService
- Simplified McpProxy to focus on MCP protocol orchestration only
- Added Microsoft.Extensions.Http package for proper HttpClient integration
